+++ src/geometry/vec2d.cpp
@@ -4,7 +4,7 @@ Vec2d::Vec2d() {
   x = y = 0.0f;
 }
 
-Vec2d::Vec2d(float x_, float y_) {
+Vec2d::Vec2d(double x_, double y_) {
   x = x_;
   y = y_;
 }
@@ -29,7 +29,7 @@ Vec2d& Vec2d::operator-=(const Vec2d& other) {
   return *this;
 }
 
-Vec2d Vec2d::operator*(const float& other) {
+Vec2d Vec2d::operator*(const double& other) {
   return Vec2d(x * other, y * other);
 }
 
@@ -39,13 +39,13 @@ Vec2d& Vec2d::operator*=(const Vec2d& other) {
   return *this;
 }
 
-Vec2d& Vec2d::operator*=(const float& other) {
+Vec2d& Vec2d::operator*=(const double& other) {
   x *= other;
   y *= other;
   return *this;
 }
 
-Vec2d& Vec2d::operator/=(const float& other) {
+Vec2d& Vec2d::operator/=(const double& other) {
   x /= other;
   y /= other;
   return *this;
